behavior_graph/lib/neo4j-jdbc-neo4j-jdbc-2.1.4/src/test/java/org/neo4j/jdbc/embedded/EmbeddedQueryExecutorTest.java
package org.neo4j.jdbc.embedded;

import org.junit.Test;

import org.neo4j.graphdb.GraphDatabaseService;
import org.neo4j.graphdb.Node;
import org.neo4j.graphdb.Transaction;
import org.neo4j.jdbc.ExecutionResult;
import org.neo4j.test.TestGraphDatabaseFactory;

import static java.util.Arrays.asList;

import static org.junit.Assert.assertEquals;

/**
 * @author mh
 * @since 15.06.12
 */
public class EmbeddedQueryExecutorTest
{
    @Test
    public void testDoExecuteQuery() throws Exception
    {
        GraphDatabaseService db = new TestGraphDatabaseFactory().newImpermanentDatabase();
        long nodeId = createNode( db );
        final EmbeddedQueryExecutor executor = new EmbeddedQueryExecutor( db );
        final ExecutionResult result = executor.executeQuery( "start n=node(" + nodeId + ") return ID(n) as id",
                null, true );
        assertEquals( asList( "id" ), result.columns() );
        final Object[] row = result.iterator().next();
        assertEquals( 1, row.length );
        assertEquals( nodeId, row[0] );
    }

    private long createNode( GraphDatabaseService db )
    {
        try ( Transaction tx = db.beginTx() )
        {
            Node node = db.createNode();
            tx.success();
            return node.getId();
        }
    }
}
